US016: added status obras cache config directory
The cache of status obras is ready to get used.

In Fact we also added an authoritative cache class
into dal.cache package of obras microservice to deal with.

In the spirit of not putting this user story off for a long time,
The current types of obra are just fake until the scrum master
shows up and feed the definitive ones.

Signed-off-by: Bob Ross <pianodaemon@gmail.com>